requirements.txt
scanpy
numpy
pandas
anndata
cellxgene-census
cellxgene-census[experimental]
leidenalg
python-igraph
scvi-tools
scipy
botocore
torch
scikit-learn
networkx